fix recent search

+++ features/digital/hotel/src/main/java/com/tokopedia/hotel/destination/view/fragment/HotelRecommendationFragment.kt
@@ -213,7 +213,7 @@ class HotelRecommendationFragment: BaseListFragment<PopularSearch, PopularSearch
     override fun onItemClicked(recentSearch: RecentSearch) {
         val intent = Intent()
         intent.putExtra(HOTEL_DESTINATION_NAME, recentSearch.property.value)
-        intent.putExtra(HOTEL_DESTINATION_ID, recentSearch.property.id)
+        intent.putExtra(HOTEL_DESTINATION_ID, recentSearch.property.id.toInt())
         intent.putExtra(HOTEL_DESTINATION_TYPE, recentSearch.property.type)
         activity?.setResult(Activity.RESULT_OK, intent)
         activity?.finish()
